Structure and Interpretation of Computer Programs
The text's version of the evaluator will need to be adapted for Racket a
little more than I expected because of the mutable pair issue (see
cchuahuico/SICP-Metacircular-Evaluator for an example of someone having
done this). Or maybe it would be possible and easier to force
classic-Scheme behavior with the `#lang` directive?