:revolving_hearts: Cracking the Coding Interview
See comment for context. This initial hack is very slow (perhaps in part
because I didn't bother implementing tries) but extremely promising. For
example, given a 4x3 grid seeded with TED and ARE in the first row (this
choice inspired from a segment of Del Puzzler's Crossword Puzzles March
2015 #1, in which 25, 33, 37, and 40 Across are COMPETED, RARE, SPIN,
and SET, respectively), it eventually (timescale of minutes) comes up
with

TED
ARE
BMW
SAY

If we change the second row of the input to APE, it eventually hits on

TED
APE
KIM
ECO

Although actually working on the crossword-generator idea hadn't made
the cut for my vague study plans for this year, I'm seriously
considering changing my mind and spinning this one out as a seprate
project.