Programming Challenges
I think today's moral lesson is that the ordering of next-move
candidates in a backtracking search is very important. The heuristic
ordering of primes-less-than-something in the summation-of-four-primes
problem gave us a huge speedup except in the cases where it doesn't (see
comment in code).

; before
found solution [2, 2, 2, 29] for 35 after 75726 calls and
1.048896312713623 seconds
found solution [2, 2, 3, 29] for 36 after 187488 calls and
2.639303207397461 seconds
found solution [2, 2, 2, 31] for 37 after 165008 calls and
2.333181619644165 seconds
found solution [2, 2, 3, 31] for 38 after 408586 calls and
6.272430658340454 seconds

; after
found solution [7, 13, 13, 2] for 35 after 714 calls and
0.016911029815673828 seconds
found solution [7, 11, 13, 5] for 36 after 6 calls and
9.870529174804688e-05 seconds
found solution [11, 13, 11, 2] for 37 after 6 calls and
9.989738464355469e-05 seconds
found solution [11, 13, 11, 3] for 38 after 6 calls and
0.00010347366333007812 seconds

I'm also quite pleased with how simple the resolution of the
fifteen-puzzle the apparent resolution of the fifteen-puzzle intrigue
turned out.