/r/dailyprogrammer
When they say type hints are in Python 3.5, do they just mean the
classes in the typing module, or is there a built-in typechecker that
actually uses them that I couldn't find? Mypy "lang" (with 3.4) is
usable, but not a joy to use (and it gave me an internalsy stack trace
rather than a proper error when I guessed wrong about Callable
annotations).

Rust is feeling more usable every week, and since we're past 1.0 (and
moreover, I haven't been running rustup.sh), it's surely because I'm
changing, not the compiler.