Stockfighter
This game is extremely demoralizing—I can write (ugly) code (that I
don't feel motivated to make less ugly when it's just to hit an API for
a game), but I can't write code that consistently makes money and I
don't know how to fix it. I technically cleared level 3 by trying out a
cheap hack someone mentioned in the
forums (https://discuss.starfighters.io/t/level-3-cheap-win-bug-or-feature/3950),
where there's some bot that makes market orders that you can exploit by
leaving a ludicrously overpriced ask and waiting until it just happens
to rest on the top of the book with the idiot market buy order comes
in. That's not a real victory—I can't expect to make progress on the
remaining levels unless I actually fix the hole in my model (and my
code).

Maybe I just need to stop being lazy—I've had this attitude that "Eh,
this isn't supposed to be hard yet, I can just bang out this sloppy
approximate code and expect it to pretty much beat the bot market." But
now I've seen that the bot market is smarter than that—only after
exerting the effort to formulate the _right_ code expressing the
_correct_ concepts, should I feel despondent and humiliated.