Stockfighter
This game is really hard! And I still think there's something wrong with
my psychology. Most of the time when I'm programming, I'm in a mindset
of trying to produce an æsthetically adequate code artifact that
incidentally happens to fulfill some functionality requirement, but when
I'm trying to "beat the level", suddenly it seems right to bang out the
most hideous, buggy crap.

I think this might be a little closer to the right track? I was blowing
my exposure limits a lot, and even now I think I'm "leaking" orders that
I don't cancel or check up on later.