/r/dailyprogrammer
I can't keep living like this. I can't. It is written (by our savior,
Alyssa Rosenbaum) that there are no evil thoughts—except one. But then
why am I alive? As the evil thought consumes entire hours and days, it's
not clear that I am in the senses that matter.

This is like the bad parts of being a teenager, before I knew that there
was such a thing as studying, before I had a tribe

I can'tsfbjkl